Added support for compile time and runtime default values for preferences

+++ app/src/main/java/com/lacronicus/easydatastore/MainActivity.java
@@ -31,11 +31,11 @@ public class MainActivity extends ActionBarActivity {
         Log.d("APP", "" + myInt);
 
         datastore.myFloat().put(2.4f);
-        float myFloat = datastore.myFloat().get(-1);
+        float myFloat = datastore.myFloat().get((float) -1);
         Log.d("APP", "" + myFloat);
 
         datastore.myLong().put(Long.MAX_VALUE);
-        long myLong = datastore.myLong().get(-1);
+        long myLong = datastore.myLong().get((long) -1);
         Log.d("APP", "" + myLong);
 
         datastore.myBoolean().put(true);
